[time-namespaced state] Move navigationOptions out of Route. (#5444)
Moves navigationOptions out of the Route.

navigationOptions currently only has a replaceState property, which is an internal detail of app_routing_effects. Nothing else in the code base depends on this value. I will be adding another property resetNamespace to navigationOptions in a followup change, which is another navigation-related internal detail of app_routing_effects.

In practice navigationOptions (replaceState and resetNamespace) is a separate concern from Route. It describes properties that we consider between possible Route transitions rather than a property of Routes themselves.