chore: add prettier-plugin-sort-imports
Prettier by default does not sort the imports because it believes in no
source code transformation (just does re-print). As such, imports were
left to the developer and causes cognitive overhead when authoring a
module since we have a soft rule about how it should be ordered.

This addition of the plugin automates the formatting of import
statements and this change applies the format onto TypeScript files.