[time-namespaced state] Support back and forward browser button. (#5451)
Support time-namespaced state for back and forward browser buttons and, really, any navigation to a point in browser history. As user navigates we store the namespace id with the history entry. When user navigates using browser history then we read the namespace id to determine which namespace to use.

Most of the changes in app_routing_effects are actually just improvements to the typing for data. We formalize the progression of data transformation -- from Navigation, to RouteMatch, to Route -- and we attach well-typed "NavigationOptions" to the data.