line chart: x-axis extent editable
After DOM structure refactor, manual extent setting feature was broken
on the x-axis only. This is due to it having `customVis` specified for
it and the customVis was taking all the pointer events, causing :hover
not to function.

This change puts `pointer-events: none` on `.custom-vis` in all cases
and also improved the CSS a little bit.

This will be tested via screenshot tests internally.