timeseries: fix setting reset causing weird behavior (#5458)
PR #5030 divided up the default setting from overrides and, in selector,
we have mixed two values in selectors to make a concrete settings value.

When reset, overriden values were set to `undefined` and caused the
settings to be very wrong.

```ts
// This resolves to {a: undefined}
{...{a: 1}, ...{a: undefined}}
```

Object spread (or `assign`) does not skip a property whose value is
`undefined` and causes settingOverride with undefined value to take
precedence over the default value which, in the end, resolves the image
brightness/contrast values to `undefined` unlike what the types
indicated.

In the end, this is a fault of the type system since the interface is
defined as `Partial<>` and a value should really not have `undefined`
value. To illustrate,

```ts
interface Foo {
  bar: number;
}

// Legal; `bar` property may not exist
const foo2: Partial<Foo> = {
};

// Legal; `bar` is defined as `number` but here, it is `undefined`
// yet it is legal.
const foo3: Partial<Foo> = {
  bar: undefined
};
```

TypeScript 4.4 has a way to property guard against `undefined` value in
a `Partial<>` type, `"exactOptionalPropertyTypes"` but TensorBoard app
currently violates the type definition so we will enable the tsconfig in
a subsequent changes.