line chart: better WebGL edge case handle (#5465)
This change makes two improvements to WebGL support for the line chart.

1. Better feature detection. We now also listen to
   `webglcontextcreationerror` which gets emitted when browser fails to
   create webgl context from a canvas. This can be a good indication
   that user's system may not be able to handle WebGL renderer.
2. Better removal of JavaScript object. When an Angular components get
   scrapped, we get a chance to clean more stuff up. With the signal, we
   now instruct Three.js renderer to dispose all the objects more
   explicitly for better freeing of WebGL context + Three.js objects.

There is no great way to unit test to guarantee correctness of the
utility and the operation. Thus, we will leave such test to an
integration test that would exercise browser capabilities.  For now, we
are happy with a manual functional test.