[time-namespaced state] Remove most usage of getRouteId(). (#5464)
We remove most usage of getRouteId() (both the internal util function and the selector).

getRouteId() has otherwise been used by the codebase to ask "are two routes the same route kind and the same set of experiments"? So we introduce a new "areSameRouteKindAndExperiments()" function to answer this question more clearly.

For the most part, getRouteId() the selector can be replaced with getActiveRoute() and possibly the new areSameRouteAndExperiments().

For the most part, getRouteId() the internal util function can be replaced with areSameRouteAndExperiments().

The final remaining usage of getRouteId() is for namespace id generation in app_routing_effects. As this form of namespace id is on the way out (googlers, see go/tb-timespaced-state), we expect to be able to remove this sometime in the new year.