(time-namespaced state) Always set namespace id in browser history state. (#5467)
Always set namespace id in browser history, no matter whether enableTimeNamespacedState is true or false.

It has no practical impact on the behavior of the application (we don't yet read this value in any case where enableTimeNamespacedState is set to false).

This is a prelimary change to fix some of the navigation issues detailed in b/211018035 (sorry, for Googlers only. I promise to be a bit more descriptive in the PR that actually fixes the issues). I made this change separate from the rest to limit the size of the PR.

In order to minimize impact on tests, a lot of which assert on the number of calls to Location.replaceState(), I refactored Location.replaceState(data, url) into two separate calls: Location.replaceStateUrl(url) and Location.replaceStateData(data). I can additionally rationalize this by pointing out that replaceState(data, url) was never called with both data and url.