(time-namespaced state) Reuse namespace id on page reload if it exists. (#5470)
This change ensures that we read and reuse namespace id from browser history if a page is reloaded.

Consider the situation where user:

1. Selects to view an experiment and sets tag filter to 'A'.
2. Adds an experiment to the comparison and then sets the tag filter to 'B'.
3. Refreshes the page.

All these actions are supposed to take place in the same namespace. But currently the page refresh breaks this - a page load always generates a new namespace and new namespace id. So if user navigates back to the single view they will see tag filter set to 'A' instead of to 'B'. This change fixes things so that when the user navigates back to the single view they will see filter set to 'B'.