time-namespaced state: Only rehydrate state from url for namespaces not seen since page load. (#5469)
As desribed in detail in b/211018035 (sorry, Googlers only), there is tension between ngrx-managed state and state as described in the URL.

We can get in situations where user navigates back and forward in browser history and the state rehydrated from the URL overwrites ngrx-managed state. In the case of namespaced state it can often overwrite the WRONG state since the rehydration happens before namespace resolution is completed.

The solution we propose is to rehydrate state from the URL less frequently. We skip rehydration from URL if navigating to a namespace that we already have tracked in ngrx-managed state.

This approach will solve a bunch of navigation bugs in route-namespaced state but will also avoid some weird potential navigation bugs in the new time-namespaced state.

It does have some certain idiosyncracies of its own that we will have to accept as intended behavior. Consider if user has generated the following browser history and namespaces since loading the app:

* (1) {http://urlA?tagFilter=A, namespaceId: A}
* (2) {http://urlAprime?tagFilter=Aprime, namespaceId: A}
* (3) {http://urlB?tagFilter=B, namespaceId: B}

Then two behaviors we'd have to accept as 'Working as Intended':

1. If user jumps from state (3) to state (1) then 'tagFilter=A' will be ignored and the tag filter will be 'Aprime'. This seems to be justifiable since it fits into our mental model about how namespaces are supposed to behave according to go/tb-timespaced-state.

2. If user reloads the page (thus clearing in-memory state) and then jumps from state (3) to state (1) then 'tagFilter=A' will be honored and the tag filter will be 'A'. If they then navigate forward to state (2) then 'tagFilter=Aprime' will be ignored and the tag filter will remain 'A'. This might be a little harder to justify but I think it is a compromise that makes sense given how the system works.