time-namespaced state: Update symbols in route_contexted_reducer_helper to use Namespace. (#5477)
Update route_contexted_reducer_helper to use "Namespace" instead of "Route" for symbols. Try to update the remainder of the code base to match. At a high level, make the following replacements in the code:

RouteContexted => NamespaceContexted
Routeful => Namespaced
Routeless => NonNamespaced

Also update the documentation in route_contexted_reducer_helper to reflect the "Namespace" concept and be less route-centric.

We do not update the file names (route_contexted_reducer_helper(_test).ts) at the moment but will in a subsequent change.

Note that this change should be safe to import into the Google-internal repo after cl/417855741 is submitted.