time-namespaced state: fix experiment data initial loading on timeNamespacedState enabled (#5508)
This only happens when the browser is initiated on experiment view and timeNamespacedState is enabled from the param. `NamespaceUpdateOption` is set to `NEW` initially and gets updated to `UNCHANGED` on any navigation afterwards.
However in loading experiment view initially there is no navigation actions therefore the option remains `NEW`, which generates a new namespaceId. 
In the fix we set the value to `UNCHANGED` after the first run of app navigation flow. However this is tricky to add the unit tests. InitAction perfectly ran only once of the changed part and trigging any follow up actions triggers navigation$.
Add webtests to catch this.