time-namespaced state: generalize navigated callback function on createNamespaceContextedState (#5511)
Instead of listening on navigated action in feature reducer, we should implement `onRouteKindOrExperimentsChanged` and pass into `createNamespaceContextedState` which then runs on route (rout kind or experiments) changed. 

However it might be unclear for reducers that the state it returns only takes place when route changed (if there is no route change, the new state is not effective)
We then move this logic out of app_routing and let feature reducers decide what they want to do on `navigated` action.

For the current metrics/ and runs/, the states are updated on route changed.