time-namespaced state: Add support to resetNamespacedState in programmatical navigation. (#5521)
Adds support to programmatical navigation to allow resetNamespacedState to be set. This is similar to how the routerLink directive supports a resetNamespacedState attribute.

If unspecified then it is treated as false (aka NamespaceUpdateOption.UNCHANGED).

I tested it by importing into the internal repo and modifying one of the programatical navigations to set this to true. I loaded the app with enableTimeNamespacedState and observed that performing the corresponding action reset the namespaced state.