Add WebGL escape hatch (#5494)
There is logic to fallback to SVG if we receive an error from WebGL. However, if some users are having trouble with WebGL and we are not catching the error we wanted an escape hatch for users to force the use of SVG.

This PR adds that escape hatch as a hidden query parameter that users can add. Once added it will be stored locally until turned off which can be done with another query parameter.