time-namespaced state: Fix hash-based navigation. (#5537)
With time-namespaced state enabled, navigation from plugin to plugin is broken. The root cause is improper handling of hash-based navigations in app_routing_effects.ts.

When a browser app navigates forward by modifying the URL hash, this navigation is signalled to the rest of the app with a popstate event. 

Up to now we had been assuming that every popstate event was the result of navigation using browser back/forward buttons. We further assumed, then, that the browser history state would have a namespace id that we could reuse in the navigated() event.

The URL hash behavior breaks these assumptions -- the browser history state does not have a namespace id. We handle this case by essentially treating it like a navigation with resetTimeNamespacedState=false. That is, we reuse the existing namespace id from the previous route.