linked time: Adds gray ticks to sliders on image cards (#5559)
The small black "tick" shown on hover in Angular `<mat-slider>` is actually created by background (with `repeating-linear-gradient`) which is not customizable. It is difficult to create dots as well as control how many ticks we want to show. 
I then created another layer on top of it to generate the gray dots. In single selection, only a single tick shows. In range selection, those ticks included in a range appears. 
There is no control over the thumb position for now.