linked time: moves Image thumb slider on single selection (#5569)
Previously we have those gray ticks to show the selected step(s) in linked time. This pr moves the thumb to the selected step, the image also changes correspondingly. There is still one tick rendered but it will be covered by orange thumb.

Note that this pr only moves the thumb (controlled by `stepIndex`) when selected step matches. Not move to the closest step. Also because of the nature of image cards (which shows one image at a time) we do not move the thumb on range selection for now before more complicated decision is made.