linked time: remove useRange input in <tb-range-input> (#5583)
When extending `<tb-range-input>` it took `useRange` to determine rendering of single or range slider. This pr further removes the explicit usage of `useRange` param but put upperValue to determine the rendering. That is to say, it is range selection when upperValue is null, otherwise it's single selection.

Summary of key changes:
* `useRange` is removed. The logic is replaced with `upperValue !== null` (equivalent to useRange is true)
* The interface`StoreInternalLinkedTime`, which does not allow end step to be null, is replaced with `LinkedTime`. 
* In reducer, on `timeSelectionChanged` action, we set `nextEndStep` to be null if the change does not have end step instead of step max.