plugins: extends plugins to take permission denied template (#5592)
Currently on data load failure the plugin page can render two different templates, "unknown" and "not found", but they are used only by tb.dev. Therefore the default template, which is showing "Data could not be loaded" without further information, is shown most of the time.
Now we want to show meaningful message on the page when the error return from server is permission denied. The plugin component now also takes the permission denied template input if provided.