Linked Time: Implement typing step into fob (#5580)
This adds the functionality of changing the selected step by typing it in directly to the fob. This is done by double clicking the fob which changes the text to an input field. Once typed the user can click enter or simply click off the Fob to change focus and the selected step will be updated. This also adds logic to allow swapping of fobs in a range selection if the step entered moves past the other Fob.