linked time: Show histogram with color on mouse hovering (#5597)
On single selection currently histogram card only show data with the matched step. We want to enable color displaying on those hovered steps.

This could not be done in css (using `:hover`) is because the color comes from d3 (with `d3ColorScale`). Thus we add/remove `.no-color` class on mouse enter/leave.