linked time: moves thumb to closest step on single selection when they're "close enough" (#5594)
Previously on single selection we only move the thumb to steps when it is exactly matched with selected step. Now we want to move the thumb (to show the corresponding image) when the selected step is "close enough".

We define the "close enough" to be 10% of the distance between two step values. For example we have steps [20, 50, 200, 400]. We move the thumb to 50 when its range is within 47-65. Calculated by 10% of (50-20) and 10% of (200-50).