linked time: decouple linked time interface from metrics to be reused in widgets (#5602)
Currently the interface `LinkedTime` defined in `metrics/internal_types` but is also used in widgets. List of usage:
* metrics/[views/stores]
* widgets/linked_time_fob
* widgets/histogram

This pr moves the definition out of metrics/ and to widgets/linked_time_fob, named "linked_time_types". The purpose of this file is for types shared with other widget/features. We will have another internal_types.ts for types only used in linked_time_fob/.