linked time: click on histogram to show the selected range (#5599)
When it's single selection and user click on other non-selected steps, we want to create a range selection between the previously selected step ("start step") and the clicked step.

I further extend the feature to be applied on range selection. That is to say, if user clicks step outside the range, we update the range by expanding the clicked step to be the new start step or end step, depends on the numbers.