linked time: fix colored histogram uncolored on mouse hovering (#5611)
Previous implement, coloring histogram on mouse hovering #5597 , cause a bug where those colored histogram will be uncolored on mouseleave. This was due to lack of checking when adding `.no-color` class. 
This pr adds a condition to check whether datum is in range. If it is in range, means we don't want to update the color on hovering. (includes removing .no-color because the class should not exist when datum is in range)