linked time: extends VisSelectedTimeClippedModule to VisSelectedTimeWarningModule (#5613)
We have a `VisSelectedTimeClippedModule` used to show small escalation icon when linked time is clipped in the cards. Now we want to show similar icon (the same icon for now, subject to change) with different title when the card shows the "closest" step instead of the exact selected step when there is no value on the selected step (think about histogram with no data on selected step)

Summary of changes:
* rename `vis_selected_time_clipped` to `vis_selected_time_warning`
* takes input `isClipped` and `isClosestStepHighlighted` to decide rendering which icon
* `isClipped` is set to true temporary for backward compatibility. It will be set to false when we roll up the changes to cards.
 
It looks unlikely a card will be both clipped and shown the closest step. But we still keep two input param for simplicity (and flexible once we need two 

Next step: Changes the rendering logic in (scalar/histogram/image) cards ng.html. Replaces `*ngIf=clipped` logic with input param (`isClipped`). The tests will be also put with these card components.