Clear tag and chart metrics data on navigated action. (#5626)
For Google-internal TensorBoards with more complex navigation, users are running into bugs where loading new experiments is not loading any chart data or loading new comparisons is not loading new chart data. (see b/225162725).

The root cause is stale tag and chart metadata that is being used to render the charts before new tag metadata is retrieved. We replace this data with the new tag metadata, anyway, so the change is to clear this data earlier, on navigated event whenever the sets of experiments in a route changes.