Create histogram fob adapter component and use with LinkedTimeFobControllerComponent (#5628)
This is a refactoring of the logic used for dragging the linked time fob in the Histogram Card. The previous implementation did not allow for the use cases needed in the Scalar Card. Our goal is to pull out the logic that will be histogram specific into a class that is generalizable. The CardFobAdapter interface created in #5623 is the generalization that will be implemented in both the Histogram and Scalar cards to allow for different uses of the draggable fob.