pip: limit werkzeug to less than 2.1 (#5642)
werkzeug package 2.1.0 is released, which BaseResponse is no longer supported in that latest version.
https://werkzeug.palletsprojects.com/en/2.0.x/wrappers/
This breaks our py tests (#5637) in both tb-nightly and CI as the env install tf-nighly with the requirement of >=0.11.15. This PR limit the werkzeug pkg version to be less than 2.1.0. This makes the syntax in current tests are still available.