time namespaced: add pinned card back when experiments are added back (#5627)
In comparison view, when an experiment is removed and added back, if there are pinned cards from these re-added experiments, we lost name even within the same namespace. We want to bring these pinned cards back when the associated experiments are added back.
This PR introduce a new state, `cardToPinnedCopyCache`, to remember all card ids that have ever been pinned and not unpinned. That is to say, we update `cardToPinnedCopy` state to know what are the pinned cards that we want to show under the current card lists. But `cardToPinnedCopyCache` is only updated on pinned/unpinned actions along with `cardToPinnedCopy`, we do not update it on `metricsTagMetadataLoaded`.