linked time: updates cardStepIndex on selected time toggled (#5641)
We've updated cardStepIndex on and only on select time changed. This causes issue when select time is enabled the very first time, there is no linked time set yet and the card step index remain unchanged. However, in settings panel we still show the min step and single selection (while the select time is null in ngrx state). This causes a bad ux when the image steps actually matches the min step: logically we should move the step index but we did not. 

This PR sets card step index when linked time is enabled. `generateNextCardStepIndexFromSelectedTime` is reused. If select time is not set, we take the min step and apply single selection to update the image card step index.

Full context please see #5615