linked time: stepIndex moving: tests for card step updated on timeSelectionChanged (#5649)
Adds tests for `generateNextCardStepIndexFromSelectedTime` usage in `metrics_reducer`. Full context please see #5615 

The changes of `selectedTime` in this reducer have been tested. The logics of the method also have been tested in `store_internal_utils` and not intent to repeat here. The tests on this higher reducer level is just to test the method returns `cardStepIndex`, which used to update (or not update) the next `cardStepIndex` state.