angular: update FormControl usages to be forward compatible with typed forms. (#5662)
TensorBoard has a couple FormControls which will be broken when typed forms lands in v14, and inside google3. Ordinarily, you'd want to use the opt-out symbol (UntypedFormControl). However, since Tensorboard needs to be simultaneously compatible with both v12 (on Github) and v14 (inside google3), we can use this cast as a trick to disable the new types.