linked time: displays indicator when thumb is moved to closest step (#5661)
In image card, when there is no data on selected step, we automatically move the thumb position to the closest step. We don't want to confused user to think about the data exist on this selected step so we add an icon indicator on the top of the card to show user something is happening.