Adjusts Input type for CustomizableComponent for better template type checking (#5677)
As part of effort to enabel strict Angular template checking in our internal code base, I need to make an adjustment to CustomizableComponent's @input type. It is now typed as Type<Component> - this is what the callers are all actually passing, so it is strictly a better type.

We still cast the constructor to Type<unknown> when passing it to the underlying ComponentFactoryResolver to satisfy the TypeScript compiler.