Unsubscribes ReloaderComponent from state changes when destroyed. (#5684)
In Google-internal TensorBoard instances, the ReloaderComponent might be loaded and destroyed multiple times. Unfortunately ReloaderComponent does not properly unsubscribe from ngrx Store changes when being destroyed. As described in b/230111530 this leads to strange behavior where we can end up having multiple auto-reload cycles in parallel because we have multiple ReloaderComponent instances (all but one is stale) trying to create these cycles.

ReloaderCompoment manually subscribes to the Store in ngOnInit(). As we've seen in other Components in our codebase this requires a special mechanism to unsubscribe (this is the takeUntil(ngUnsubscribe) line). We reuse the pattern used in the other Components and apply it here.