linked time: refactor fob step editing with contenteditable (#5687)
Previously we can edit the step within the fob with the help of adding an input element. The actions were double clicking the div and then click to focus the input. This PR refactor removes input and directly use contenteditable in span element, which allows user to click the number and edit it right away.