linked time: use 'safe' way of setting innerHTML (#5695)
When dragging the fob, we want the number in the fob to reflect the step change. With the change of contenteditable, we need to change the html of the span element. Previously we grab the element and change ".innerText" directly. I didn't check and do any kind of sanitization at all. That was a mistake. This PR use "[innerHTML]" which Angular will do the sanitization work.