Time Series: Elevate "Time Series" dashboard to first dashboard. (#5698)
The "Time Series" dashboard is a unified view of scalars, histograms, and images data. It has been available for some time but is generally the last dashboard listed in the TensorBoard header. After quite a bit of work improving it and using it internally at Google, we've decided it is time to elevate it to be the first dashboard.

Please find more information or feel free to leave feedback at #5704.