Adds mechanism to include feature flags in Angular HTTP requests. (#5718)
This adds a mechanism that dumps the client-side FeatureFlag object into each HTTP request that originates from the TB Angular code base.

(Googlers, see: http://go/tb-client-feature-flags-in-data-provider for the motivation and high level design.)

The data is dumped into a custom header. The format is effectively:
X-TensorBoard-Feature-Flags: JSON.stringify(currentFeatureFlags)

We implement this by introducing an HttpInterceptor. It allows us to intercept each request in the Angular HTTP pipeline and inject the new header. The HttpInterceptor is provided in the existing FeatureFlagModule so all instances of TensorBoard will receive this behavior without any additional configuration.

Additional things worth noting:

* This will send ALL feature flags in the request header rather than a curated subset of them. In a subsequent PR, we'll add support to mark certain feature flags as "sendToServer" and only send that subset. This PR unblocks the main use case at the cost of there being some temporary overhead in HTTP requests.

* In a subsequent PR We will add similar functionality to the TB Polymer code base.

* In a subsequent PR we will add support on the TB HTTP Server to consume these feature flags and trigger logic based on the values.

* I moved feature_flags tests into their own "karma_test" target separate from the big top-level karma_test target. stephanwlee advocated for breaking up the top-level karma_test target in the months before he left the team.