Adds client_feature_flags to DataProvider's RequestContext. (#5721)
Add a client_feature_flags property to the DataProvider's RequestContext object. Also add a middleware that extracts the feature flags from the HTTP request and sets them on the RequestContext.

client_feature_flags represents feature flags set in the TB webapp client. They will be injected into many HTTP requests originating from the webapp and will be of the form:
X-TensorBoard-Feature-Flags: JSON.stringify(currentFeatureFlags)

We expose these values as a Dict in the RequestContext object, with string keys and values of arbitrary types (but, in practice, mostly booleans and sometimes arrays of strings). Usage of client_feature_flags should be aware of the key names and the value types.

(Googlers, see: http://go/tb-client-feature-flags-in-data-provider for motivation, more details, and high level design.)

#5718 and #5727 are complimentary PRs that set the feature flags in requests originating from the Angular and Polymer code bases, respectively.