Validate that decoded client_feature_flags is a dict. (#5728)
In https://github.com/tensorflow/tensorboard/pull/5721 we added ClientFeatureFlagsMiddleware. It decodes the feature flag payload from JSON into a dict.

Reviewer suggested we validate that the decoded result is indeed a dict and I claimed that this was not a valid scenario to worry about. I realized not long afterwards that I was wrong.