CI: Add nightly release workflow
Adds the `nightly-release` workflow, which runs the entire CI workflow
before attempting to upload the `tb-nightly` pip package. For now, this
is set up to upload to testpypi.python.org so that we can validate this
process.

Additionally, updates the permissions on `update_version.sh` to avoid
needing to `chmod` it within the workflow itself.