Make grpc.patch instructions more accurate. (#5795)
As noted in a comment for #5793, the instructions for generating grpc.patch would be more accurate if they referred to the commit that corresponds to the grpc archive instead of referring to the nearest tag. I've updated the instructions to refer to the commit.

I tested by following the updated instructions and copying grpc.patch to my tensorboard repo. git did not find a diff in the generated files.